# Change EncodeBase64 to return char array