I'll start with some easy wins that benefit all targets. We'll see what fun we can do with Span later.

I have changed the string interpolation back to `StringBuilder` but presized them to fit the return contents exactly.

I changed the `EncodeBase64` method to return a char array for which I precalculate the size the encoded text will have. These can be put directly into the `StringBuilder`.

GenerateSalt performance almost doubles, and slightly improves memory usage.
(On master GenerateSalt requires 680 bytes, so definitely improving already)

Before:

|              Method |                 text |                 hash |                value |             Mean |         Error |        StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------------- |--------------------- |--------------------- |-----------------:|--------------:|--------------:|-------:|------:|------:|----------:|
|        **GenerateSalt** |                    **?** |                    **?** |                    **?** |         **392.7 ns** |       **0.87 ns** |       **0.81 ns** | **0.1125** |     **-** |     **-** |     **472 B** |
| PasswordNeedsRehash |                    ? |                    ? |                    ? |       1,615.3 ns |       2.47 ns |       2.07 ns | 0.3166 |     - |     - |    1328 B |
|      **VerifyPassword** |                     **** | **$2a$1(...)QPlxO [60]** |                    **?** | **250,126,821.4 ns** | **541,818.72 ns** | **480,308.18 ns** |      **-** |     **-** |     **-** |  **268628 B** |
|        **HashPassword** |                    **?** |                    **?** |                     **** |   **3,974,860.5 ns** |   **1,750.23 ns** |   **1,461.52 ns** |      **-** |     **-** |     **-** |   **10232 B** |
|      **VerifyPassword** | **abcde(...)vwxyz [26]** | **$2a$1(...)Qk7dq [60]** |                    **?** |  **62,594,599.0 ns** |  **19,481.85 ns** |  **15,210.15 ns** |      **-** |     **-** |     **-** |   **71548 B** |
|        **HashPassword** |                    **?** |                    **?** | **abcde(...)vwxyz [26]** |   **3,975,397.4 ns** |   **1,227.28 ns** |   **1,024.83 ns** |      **-** |     **-** |     **-** |   **10342 B** |

After:

|              Method |                 text |                 hash |                value |             Mean |        Error |       StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------------- |--------------------- |--------------------- |-----------------:|-------------:|-------------:|-------:|------:|------:|----------:|
|        **GenerateSalt** |                    **?** |                    **?** |                    **?** |         **196.1 ns** |      **0.50 ns** |      **0.44 ns** | **0.0861** |     **-** |     **-** |     **360 B** |
| PasswordNeedsRehash |                    ? |                    ? |                    ? |       1,644.3 ns |      1.94 ns |      1.62 ns | 0.3166 |     - |     - |    1328 B |
|      **VerifyPassword** |                     **** | **$2a$1(...)QPlxO [60]** |                    **?** | **254,804,804.2 ns** | **67,121.11 ns** | **52,403.74 ns** |      **-** |     **-** |     **-** |  **268268 B** |
|        **HashPassword** |                    **?** |                    **?** |                     **** |   **4,080,080.2 ns** | **48,618.23 ns** | **45,477.52 ns** |      **-** |     **-** |     **-** |    **9886 B** |
|      **VerifyPassword** | **abcde(...)vwxyz [26]** | **$2a$1(...)Qk7dq [60]** |                    **?** |  **63,837,940.0 ns** | **83,111.51 ns** | **77,742.56 ns** |      **-** |     **-** |     **-** |   **71195 B** |
|        **HashPassword** |                    **?** |                    **?** | **abcde(...)vwxyz [26]** |   **3,978,107.2 ns** |  **3,016.85 ns** |  **2,519.20 ns** |      **-** |     **-** |     **-** |    **9995 B** |