@jvandertil 2020-01-29
It seems I pasted the wrong benchmark into the PR, updated.