@ChrisMcKee 2020-01-30
just to its all linked https://github.com/BcryptNet/bcrypt.net/issues/54